Removed problematic library.
Instead of using Float type, switched them to double so we don't depend on a problematic library.

+++ lib/logic/Plato.dart
@@ -1,10 +1,10 @@
-import 'dart:ffi';
+//import 'dart:ffi';
 
 class Plato {
 
   String idPlato;
   String descripcion;
-  Float precio;
+  double precio;
   int stock;
   String imagen;
